Factor out image endpoint into own controller
