+++ src/Client/Pages/Admin/ContractForm.razor
@@ -75,7 +75,7 @@
 
         Console.WriteLine(JsonConvert.SerializeObject(_contract, Formatting.Indented));
 
-        HttpResponseMessage response = await _http.PostAsync("/api/v1/Contracts/new/image", _content);
+        HttpResponseMessage response = await _http.PostAsync("/api/v1/images", _content);
 
         if (response.IsSuccessStatusCode)
         {
