+++ src/Server/Controllers/ContractsController.cs
@@ -1,6 +1,5 @@
 ﻿using Application.Contracts;
 using Application.Exceptions;
-using Application.Images;
 
 using Domain.Contracts;
 
@@ -13,23 +12,19 @@ namespace Server.Controllers;
 /// </summary>
 [ApiController]
 [Route("api/v1/[controller]")]
-public class ContractsController : Controller
+public class ContractsController : BaseApiController<ContractsController>
 {
     private readonly IContractService _contracts;
-    private readonly IImageRepository _images;
-    private readonly ILogger<ContractsController> _logger;
 
     /// <summary>
     /// Constructs contract API.
     /// </summary>
-    /// <param name="contracts">The contract logic.</param>
-    /// <param name="images">The place to store images.</param>
     /// <param name="logger">The logging provider.</param>
-    public ContractsController(IContractService contracts, IImageRepository images, ILogger<ContractsController> logger)
+    /// <param name="contracts">The contract logic.</param>
+    public ContractsController(ILogger<ContractsController> logger, IContractService contracts)
+        : base(logger)
     {
         _contracts = contracts;
-        _images = images;
-        _logger = logger;
     }
 
     /// <summary>
@@ -40,31 +35,6 @@ public class ContractsController : Controller
     public IEnumerable<Contract> AllContracts()
     {
         return _contracts.FetchAllContracts();
-    }
-
-    /// <summary>
-    /// Uploads a new contract image file.
-    /// </summary>
-    /// <returns>The identifier of the stored image.</returns>
-    /// <response code="400">The uploaded file is not a valid image.</response>
-    [HttpPost("new/image")]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    public async Task<ActionResult<string>> UploadImageAsync()
-    {
-        IFormFile file = Request.Form.Files[0];
-        if (file is null)
-            throw new ArgumentNullException(nameof(file));
-
-        _logger.LogInformation("Trying to upload an image file: {Name}", file.Name);
-        try
-        {
-            return await _images.StoreAsync(file.OpenReadStream()).ConfigureAwait(false);
-        }
-        catch (InvalidImageException e)
-        {
-            _logger.LogInformation("Error occured during image upload: {Error}", e.Message);
-            return BadRequest();
-        }
     }
 
     /// <summary>
@@ -83,7 +53,7 @@ public class ContractsController : Controller
         }
         catch (IdentifierAlreadyTakenException e)
         {
-            _logger.LogInformation("ID of contract was already taken: {Error}", e.Message);
+            Logger.LogInformation("ID of contract was already taken: {Error}", e.Message);
             return BadRequest();
         }
